Added to methods that compute two functions
The first method computes the forward propagation by computing the weighted some of each node.

Then the second method uses the sigmoid function which I read and found the function for in the link in the javadoc comments above the method.
This method should return a number between 0 and 1. There are some good visual representation of what this looks like in a neural network.

Overall this main class returns the weighted sum and output after the sigmoid function.