Added more methods
Figuring out train method was very hard spent a lot of time figuring this out with all of the formulas and how to represent it in code was very hard. Spent a couple of hours on this method. Now I have to actually use it now in my main method which Idk how long that will take me honestly. But I find it interesting all the math calculations are. doing this for a single neuron because I feel like multiple neurons will be more and more difficult.